Forgot Password implented
Forgot password page implemented as a dialog box that shows up on the
welcome page.

Note: The dialog doesn’t trigger as it should. I know why, but I have
not found a workaround for it yet. The obvious solution is to place the
PasswordForget component within the login form, but that causes the
onSubmit button on the ForgottenPassword to trigger the form of its
parent as well, i.e. the login form…